work

+++ PleasePass/server/services/CommentService.js
@@ -1,3 +1,4 @@
+import { Forbidden } from "@bcwdev/auth0provider/lib/Errors"
 import { dbContext } from "../db/DbContext"
 
 
@@ -8,10 +9,17 @@ class CommentService {
   async deleteComment(id, userId) {
     const comment = await dbContext.Comments.findById(id)
     // @ts-ignore
-
-    await comment.delete()
+    if (comment.creatorId.toString() != userId) {
+      throw new Forbidden("'You can't delete that")
+    }
     // @ts-ignore
-    await comment.save()
+    const towerEvent = await dbContext.
+      TowerEvents.findById(comment.eventId)
+    // @ts-ignore
+    await towerEvent.save()
+    // @ts-ignore
+    await comment.delete()
+    return comment
   }
   async getEventComments(eventId) {
     const comments = await dbContext.Comments.find({ eventId }).populate('creator', 'name picture')
@@ -19,7 +27,17 @@ class CommentService {
     return comments
 
 
-
+    //   async remove(collabId, userId) {
+    //             const collab = await dbContext.Collaborators.findById(collabId)
+    //             if (!collab) {
+    //                 throw new BadRequest('no collab at that id')
+    //             }
+    //             if (collab.accountId.toString() != userId) {
+    //                 throw new Forbidden('You can not remove that')
+    //             }
+    //             await collab.remove()
+    //             return 'collab ended'
+    //         }
 
 
     //  async getEventTickets(eventId) {
@@ -34,6 +52,14 @@ class CommentService {
     return comment
   }
 
+
+  async getCommentById(commentId) {
+    let comment = await dbContext.Comments.findById(commentId)
+    if (!comment) {
+      throw new Forbidden('Invalid Comment Id')
+    }
+    return comment
+  }
 }
